Complete application logic

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,18 @@
 import AsyncStorage from "@react-native-async-storage/async-storage"
 import { GROUP_COLLECTION } from "@storage/storageConfig"
+import { AppError } from "@utils/AppError"
 import { groupGetAll } from "./groupGetAll"
 
 export async function groupCreate(newGroup: string){
     try {
         const storedGroups = await groupGetAll()
+
+        const groupAlreadyExists = storedGroups.includes(newGroup)
+
+        if (groupAlreadyExists) {
+            throw new AppError("There is already a group registered with that name!")
+        }
+
         const storage = JSON.stringify([...storedGroups, newGroup])
 
         await AsyncStorage.setItem(GROUP_COLLECTION, storage)
